Values are very close to being correct.
Variance passes the float epsilong value.

+++ src/choreograph/Sequence.hpp
@@ -207,7 +207,7 @@ public:
   float getTimeWrapped( float time, float inflectionPoint = 0.0f ) const
   {
     if( time > getDuration() ) {
-      return inflectionPoint + std::fmod( time, getDuration() - inflectionPoint );
+      return inflectionPoint + std::fmodf( time, getDuration() - inflectionPoint );
     }
     else {
       return time;
